added the error mod but didnt use it yet, added testing, curently testing /users GET and /users POST, split the models for user stuff and routes into seperate folders, didnt touch elm
