+++ server-axum/Cargo.toml
@@ -9,7 +9,13 @@ edition = "2021"
 axum = "0.6.15"
 rand = "0.8.5"
 serde = { version = "1.0.160", features = ["derive"] }
+serde_json = "1"
 tokio = { version = "1.27.0", features = ["full"] }
 # Tower is needed for the middleware we use (in particular, adding shared state)
 tower = "0.4.13"
 tower-http = { version = "0.4.0", features = ["add-extension", "cors"] }
+
+
+[dev-dependencies]
+anyhow = "1"
+httpc-test = "0.1.1"
